Tests for ::target-text.
These tests were written when ::target-text was implemented, but
changing window.location.hash for target text was not allowed at that
time, and did not allow for automated testing. Submit since that is now
possible.

target-text-004.html fails because currentColor for ::selection does not
fall back to using the text color from ::target-text.

Bug: 1136817, 1147859
Change-Id: Id512caec6ef7df8380f75fe82c79af27975991d1