Need to propagate tallest unbreakable before forced breaks.
Even if there's a forced break inside the child added to the fragment
builder, we need to propagate the tallest unbreakable block-size, to get
column balancing right.

This fixes an existing test, but I felt that we needed a test that more
clearly illustrates this fix, so added that. The existing test still
fails in legacy layout, but for a different reason (layout is correct,
but outline painting order is wrong).

Took the opportunity to rename PropagateBreak() to PropagateBreakInfo(),
since this method doesn't only care about actual breaks. Also, no need
to call this one if no block fragmentation is involved.

Bug: 829028
Change-Id: Ie951817095267567b9fae578fd519b4ba76dc24c